consolidates theming detail into one file
